% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{ltm_mcmc}
\alias{ltm_mcmc}
\title{MCMC LTM}
\usage{
ltm_mcmc(x, y, burnin = 2000, iter = 8000, K = 3)
}
\arguments{
\item{x}{data points}

\item{y}{response variable}

\item{burnin}{number of burnin iterations}

\item{iter}{number of iterations after burnin}

\item{K}{parameter K}
}
\description{
Given \code{x} and \code{y} performs the MCMC optimization.
}
