% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{ltm_sim}
\alias{ltm_sim}
\title{Simulate LTM model}
\usage{
ltm_sim(ns, nk, ni, vmu, mPhi, mSigs, dsig, vd, alpha)
}
\arguments{
\item{ns}{number of times}

\item{nk}{number of covariates}

\item{ni}{number of series}

\item{vmu}{vector mu}

\item{mPhi}{phi matrix}

\item{mSigs}{sigma eta vector}

\item{dsig}{general sigma}

\item{vd}{threshold parameter}

\item{alpha}{intercept}
}
\value{
list containing the generated y, x, beta and thresholded beta
}
\description{
Simulate LTM model
}
