% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot_betas}
\alias{plot_betas}
\title{Plot adjusted betas}
\usage{
plot_betas(a, betas = "all", scales = "free_y", real_values = NULL)
}
\arguments{
\item{a}{result of \code{\link{ltm_mcmc}},}

\item{betas}{which betas (indices) to plot. Defaults to all betas.}

\item{scales}{should we use free scales? Like in \code{ggplot2} package.}

\item{real_values}{simulated data from \code{\link{ltm_sim}} to compare, if it exists.}
}
\description{
Plot adjusted betas to easily compare with simulated data
}
